[REFACT] : moved ProtectedModal from Rooms.tsx to Room.tsx - 특정 room에 대한 모달이니 의미상 Room.tsx에 묶어놓는게 어울린다고 판단
